test half

+++ src/cuda/gemm_tensorcores/matrixMul/matrixMul.cu
@@ -306,7 +306,7 @@ int main(int argc, char **argv) {
          dimsB.x, dimsB.y);
 
 
-  int matrix_result = MatrixMultiply<__half>(argc, argv, block_size, dimsA, dimsB);
+  int matrix_result = MatrixMultiply<float>(argc, argv, block_size, dimsA, dimsB);
 
   exit(matrix_result);
 }
